requirements.txt
streamlit==1.52.2
python-dotenv==1.2.1
pillow==10.1.0
requests==2.32.0
httpx>=0.26,<0.29
supabase==2.27.1
langchain==1.2.3
langgraph==1.0.6
langchain-google-genai==4.1.3
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
openai>=1.8.5
pydantic>=2.7.4,<3.0.0
